+++ moyopy/src/lib.rs
@@ -5,6 +5,7 @@ pub mod base;
 pub mod data;
 pub mod dataset;
 pub mod identify;
+pub mod subgroup;
 
 use crate::base::{
     PyCollinearMagneticCell, PyMagneticOperations, PyMoyoError, PyNonCollinearMagneticCell,
@@ -22,6 +23,10 @@ use crate::dataset::{
 };
 use crate::identify::{
     PyLayerGroup, PyMagneticSpaceGroup, PyPointGroup, PySpaceGroup, integral_normalizer,
+};
+use crate::subgroup::{
+    PyTranslationengleicheSubgroup, PyTranslationengleicheSubgroupConjugacyClass,
+    PyTranslationengleicheSubgroupConjugate, enumerate_translationengleiche_subgroups,
 };
 
 // https://github.com/pydantic/pydantic-core/blob/main/src/lib.rs
@@ -89,6 +94,12 @@ fn moyopy(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<PyLayerGroup>()?;
     m.add_class::<PyMagneticSpaceGroup>()?;
     m.add_wrapped(wrap_pyfunction!(integral_normalizer))?;
+
+    // subgroup
+    m.add_class::<PyTranslationengleicheSubgroup>()?;
+    m.add_class::<PyTranslationengleicheSubgroupConjugate>()?;
+    m.add_class::<PyTranslationengleicheSubgroupConjugacyClass>()?;
+    m.add_wrapped(wrap_pyfunction!(enumerate_translationengleiche_subgroups))?;
 
     Ok(())
 }